# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Card Management System API
  description: REST API для системы управления банковскими картами с аутентификацией JWT
  version: 1.0.0
  contact:
    name: Bank API Support
    email: support@bank.com

servers:
  - url: http://localhost:8080/api
    description: Development server

tags:
  - name: Authentication
    description: Аутентификация и регистрация пользователей
  - name: Cards
    description: Управление банковскими картами
  - name: Transfers
    description: Переводы между картами
  - name: Admin
    description: Административные функции

paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: Вход в систему
      description: Аутентификация пользователя и получение JWT токена
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
        '400':
          description: Неверные учетные данные
        '401':
          description: Неавторизованный доступ

  /auth/register:
    post:
      tags: [Authentication]
      summary: Регистрация пользователя
      description: Создание нового пользователя с ролью USER
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Пользователь успешно зарегистрирован
        '400':
          description: Пользователь с таким именем или email уже существует

  /cards:
    get:
      tags: [Cards]
      summary: Получить карты пользователя
      description: Возвращает список карт текущего пользователя с пагинацией и фильтрацией по статусу
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
          description: Номер страницы
        - name: size
          in: query
          schema:
            type: integer
            default: 20
          description: Размер страницы
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/CardStatus'
          description: Фильтр по статусу карты
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/CardDto'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
        '401':
          description: Неавторизованный доступ

  /cards/{cardId}:
    get:
      tags: [Cards]
      summary: Получить информацию о карте
      description: Возвращает детальную информацию о конкретной карте
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
          description: ID карты
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDto'
        '404':
          description: Карта не найдена
        '403':
          description: Доступ запрещен

  /cards/{cardId}/block:
    put:
      tags: [Cards]
      summary: Заблокировать карту
      description: Блокировка карты пользователем
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Карта успешно заблокирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDto'
        '404':
          description: Карта не найдена
        '403':
          description: Доступ запрещен

  /cards/{cardId}/activate:
    put:
      tags: [Cards]
      summary: Активировать карту
      description: Активация заблокированной карты
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Карта успешно активирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDto'
        '404':
          description: Карта не найдена
        '403':
          description: Доступ запрещен

  /transfers:
    post:
      tags: [Transfers]
      summary: Перевод между картами
      description: Перевод средств между картами одного пользователя
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Перевод выполнен успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '400':
          description: Ошибка валидации или недостаточно средств
        '403':
          description: Доступ запрещен

  /admin/cards:
    get:
      tags: [Admin]
      summary: Получить все карты (админ)
      description: Возвращает список всех карт в системе с пагинацией
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
          description: Номер страницы
        - name: size
          in: query
          schema:
            type: integer
            default: 20
          description: Размер страницы
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/CardDto'
        '403':
          description: Недостаточно прав

    post:
      tags: [Admin]
      summary: Создать карту (админ)
      description: Создание новой карты для пользователя
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardCreateRequest'
      responses:
        '200':
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDto'
        '403':
          description: Недостаточно прав
        '404':
          description: Пользователь не найден

  /admin/cards/{cardId}:
    delete:
      tags: [Admin]
      summary: Удалить карту (админ)
      description: Удаление карты из системы
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Карта успешно удалена
        '403':
          description: Недостаточно прав
        '404':
          description: Карта не найдена

  /admin/cards/{cardId}/status:
    put:
      tags: [Admin]
      summary: Обновить статус карты (админ)
      description: Изменение статуса карты администратором
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardUpdateRequest'
      responses:
        '200':
          description: Статус карты обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDto'
        '403':
          description: Недостаточно прав
        '404':
          description: Карта не найдена

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "user1"
        password:
          type: string
          format: password
          example: "password123"

    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: "newuser"
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 6
          example: "securepassword"

    JwtResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT токен
        type:
          type: string
          default: "Bearer"
        username:
          type: string
        role:
          type: string

    CardDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cardNumber:
          type: string
          description: Маскированный номер карты
          example: "**** **** **** 1234"
        cardHolder:
          type: string
          example: "ИВАН ИВАНОВ"
        expiryDate:
          type: string
          format: date
          example: "2026-12-31"
        status:
          $ref: '#/components/schemas/CardStatus'
        balance:
          type: number
          format: double
          example: 1500.50
        userId:
          type: integer
          format: int64
        username:
          type: string

    CardCreateRequest:
      type: object
      required:
        - userId
        - cardHolder
      properties:
        userId:
          type: integer
          format: int64
          example: 1
        cardHolder:
          type: string
          example: "ИВАН ИВАНОВ"

    CardUpdateRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/CardStatus'

    TransferRequest:
      type: object
      required:
        - fromCardId
        - toCardId
        - amount
      properties:
        fromCardId:
          type: integer
          format: int64
          example: 1
        toCardId:
          type: integer
          format: int64
          example: 2
        amount:
          type: number
          format: double
          minimum: 0.01
          example: 100.00

    TransferResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number
          format: double
        status:
          type: string
          example: "COMPLETED"
        timestamp:
          type: string
          format: date-time

    CardStatus:
      type: string
      enum: [ACTIVE, BLOCKED, EXPIRED]
      description: |
        ACTIVE - Карта активна
        BLOCKED - Карта заблокирована  
        EXPIRED - Срок действия истек